models/projectMatcher/requirements.txt
flask
pandas
scikit-learn
joblib
mysql-connector-python